requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6
google-generativeai>=0.8.0
pydantic==2.11.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.12
